Remove geospatial function from `land-boundary-context.js`
Note that we have renamed this from `geospatial-or-land-boundary-context.js` to reflect the fact we no longer have any geospational function in it
